# docs/api/openapi/python-api.openapi.json
{
  "components": {
    "schemas": {
...
      "Body_ocr_extract_api_ocr_extract_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          }
        },
        "title": "Body_ocr_extract_api_ocr_extract_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "\nPython backend for Smart Ration HSD2C (side-by-side migration from the C#/.NET API).\n\n* Versioned: every route below answers under `/api/v1/...` as well as `/api/...` (same behaviour).\n* Every response uses the envelope `{success, message, data, errors}`; failures add `errorCode`.\n* `/api/*` routes that are not yet migrated are transparently served by the C# API\n  (fallback proxy); those do not appear in this document yet.\n",
    "title": "Smart Ration HSD2C API (Python)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/ai/alerts/active": {
      "get": {
        "operationId": "alerts_active_api_ai_alerts_active_get",
        "parameters": [
          {
            "in": "query",
            "name": "shopId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shopid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Open + under-review alerts (runs a throttled refresh first)",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/alerts/list": {
      "get": {
        "operationId": "alerts_list_api_ai_alerts_list_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "shopId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shopid"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Alerts (filterable by status / shop / source)",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/alerts/shop/{shop_id}": {
      "get": {
        "operationId": "alerts_for_shop_api_ai_alerts_shop__shop_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "title": "Shop Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "active",
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One shop's alerts",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/api/ai/alerts/{alert_id}": {
      "get": {
        "operationId": "alert_detail_api_ai_alerts__alert_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One alert",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/alerts/{alert_id}/resolve": {
      "post": {
        "operationId": "alert_resolve_api_ai_alerts__alert_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "alert_id",
            "required": true,
            "schema": {
              "title": "Alert Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mark an alert under review / resolved / dismissed (officials)",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/analytics/forecast": {
      "get": {
        "operationId": "analytics_forecast_api_ai_analytics_forecast_get",
        "parameters": [
          {
            "in": "query",
            "name": "shopId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shopid"
            }
          },
          {
            "in": "query",
            "name": "horizonDays",
            "required": false,
            "schema": {
              "default": 7,
              "title": "Horizondays",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "en",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "AI demand forecast (falls back to the built-in trend)",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/analytics/inventory": {
      "get": {
        "operationId": "analytics_inventory_api_ai_analytics_inventory_get",
        "parameters": [
          {
            "in": "query",
            "name": "shopId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shopid"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "en",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "AI inventory analytics (falls back to built-in stock risk)",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/analytics/queue": {
      "get": {
        "operationId": "analytics_queue_api_ai_analytics_queue_get",
        "parameters": [
          {
            "in": "query",
            "name": "shopId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shopid"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "en",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "AI queue analytics (falls back to the built-in estimate)",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/analytics/risk": {
      "get": {
        "operationId": "analytics_risk_api_ai_analytics_risk_get",
        "parameters": [
          {
            "in": "query",
            "name": "shopId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shopid"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "en",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "AI beneficiary risk ranking (officials)",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/analytics/shops": {
      "get": {
        "operationId": "analytics_shops_api_ai_analytics_shops_get",
        "parameters": [
          {
            "in": "query",
            "name": "shopId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shopid"
            }
          },
          {
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "default": "en",
              "title": "Lang",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "AI shop monitor",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/beneficiaries/{beneficiary_id}/insight": {
      "get": {
        "operationId": "beneficiary_insight_api_ai_beneficiaries__beneficiary_id__insight_get",
        "parameters": [
          {
            "in": "path",
            "name": "beneficiary_id",
            "required": true,
            "schema": {
              "title": "Beneficiary Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One beneficiary's risk summary",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/demand-forecast": {
      "get": {
        "operationId": "demand_forecast_api_ai_demand_forecast_get",
        "parameters": [
          {
            "in": "query",
            "name": "shopId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shopid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trailing-30-day demand trend",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/api/ai/inventory-risk": {
      "get": {
        "operationId": "inventory_risk_api_ai_inventory_risk_get",
        "parameters": [
          {
            "in": "query",
            "name": "shopId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shopid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stock-cover risk per shop and item",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/queue-prediction": {
      "get": {
        "operationId": "queue_prediction_api_ai_queue_prediction_get",
        "parameters": [
          {
            "in": "query",
            "name": "shopId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shopid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Today's queue wait estimate",
        "tags": [
          "ai"
        ]
      }
    },
    "/api/ai/shops/{shop_id}/insight": {
      "get": {
        "operationId": "shop_insight_api_ai_shops__shop_id__insight_get",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "title": "Shop Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One shop's risk summary",
        "tags": [
          "ai"
        ]
      }
    },
...
    "/api/inventory": {
      "get": {
        "operationId": "inventory_list_api_inventory_get",
        "parameters": [
          {
            "in": "query",
            "name": "shopId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Shopid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stock lines (my shop, or any shop for officials)",
        "tags": [
          "shop counter"
        ]
      },
      "post": {
        "operationId": "inventory_create_api_inventory_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Add a stock line to a shop",
        "tags": [
          "shop counter"
        ]
      }
    },
...
    "/api/ocr/extract": {
      "post": {
        "operationId": "ocr_extract_api_ocr_extract_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ocr_extract_api_ocr_extract_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suggest fields from a ration-card photo (PNG/JPEG, 5 MB; nothing stored)",
        "tags": [
          "ai"
        ]
      }
    },
...
  }
}